client/requirements.txt
colorama
pycodestyle
pre-commit
cutie
tabulate
requests
python-dotenv